Show tags in other channels profile field to make it easier to navigate to the channels
